% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomdp_solve.R
\name{pomdp_solve}
\alias{pomdp_solve}
\alias{pomdp_solve_help}
\title{Solving a POMDP with 'pomdp-solve'}
\usage{
pomdp_solve(pomdp, options = list(), verbose = TRUE)

pomdp_solve_help()
}
\arguments{
\item{pomdp}{the POMDP file to solve.}

\item{options}{a list with options for pomdp-solve.}

\item{verbose}{logical; show the program text output?}
}
\value{
nothing
}
\description{
This function provides a bare bones interface to run pomdp-solve on a POMDP
file. The results can be read with the function provides in \link{read_write}.
}
\details{
The options are specified in \code{options} as a list with entries of the form \verb{<option> = <value>}.
\code{pomdp_solve_help()} displays the available options. Note that the leading dash is not used on the option name. For example:
\code{list(method = "grid", epsilon = 0.1)} sets the method option to grid and epsilon to 0.1.
Here is a slightly more \href{https://mhahsler.github.io/pomdpSolve/pomdp-solve_manual}{detailed description of pomdp-solve's options.}
}
\examples{
# display available arguments
pomdp_solve_help()

# solve a POMDP file that ships with this package in a temporary directory
tmp <- tempdir()
old_dir <- setwd(tmp)

file.copy(system.file("tiger.aaai.POMDP", package = "pomdpSolve"), "./tiger.aaai.POMDP")

pomdp_solve("tiger.aaai.POMDP", options = list(method = "grid"))
dir()

# read the raw policy graph (-0 means infinite horizon solution)
read_pg_file("tiger.aaai-0.pg")

# read the raw value function
read_alpha_file("tiger.aaai-0.alpha")

# return to the old directory
setwd(old_dir)
}
\references{
Anthony R. Cassandra, pomdp-solve documentation,
\url{https://www.pomdp.org/code/index.html}
}
\seealso{
find_pomdp_solve read_write
}
